requirements.txt
panasonic_viera
aiohttp
paho-mqtt
